add visualizations for basic blocks and function arguments
Three new visualizations are introduced in this commit: basicblocks,
funcargs, and treemap.

The basicblocks visualization is a sunburst graph that contains
run-times (or number of executions) of functions with their basic
blocks. It supports basic mapping of the basic blocks to the source
code. The mapping is, however bit tricky to incorporate into the graph,
thus treemap visualization utilizing different library was created.
This visualization uses old version of bokehlibrary and might be retired
in the future.

The funcargs visualization is a scatter graph that shows dependency
between function run-time and values of collected arguments.

The treemap visualization has multiple forms of representation as a
graph: tree-map, sunburst, and flame graph. Each visualizes
the call graph as a tree in a slightly different way. It supports
mapping of basic blocks to source code.