add temporary fix for incompatibility of pin with g++ 12 and newer
There is a variable in the standard library of C++ within PIN kit that
causes uninitialized warning to trigger when using g++ 12 and newer.
Since the kit treats warnings as errors by default, the pragma to ignore
this warning was added until the release of new PIN kit with this issue
fixed is released.